Refactoring of logging to M-Files object
The Serilog.Sinks.MFilesObject sink cannot be used in the VaultApplication. We don't control the vault reference lifetime and the underlying COM object is released before first use in the logger. An error "COM object that has been separated from its underlying RCW cannot be used." throws.
Thus a DelegatingTextSink is added that makes the Vault Application buffer log messages; a background task flushes the buffered log messages to an MFiles Log object (re-using the Serilog.Sinks.MFilesObject  EmitToMFilesLogObject() function.
Initializing a flushlogAction delegate in StartApplication and reusing this in the background operation; also use this to flush in UninitializeApplication
Adding Serilog.Sinks.MFilesSysUtilsEventLog, which logs to the Event log using SysKit