feat: new implementation of using MFilesLogObjectMessage and MFilesLogFile sinks, with lazy initialization of the logger with PermanentVault
A new implementation of using the MFilesLogObjectMessage and MFilesLogFile using lazy initialization of the logger which uses the PermanentVault to batch log messages to the vault. It seems to work.
Remove the previous solution where the log messages where buffered using the DelegatingTextSink and a delegate Action() to flush it to the M-Files vault. Preserved the solution in "VaultApplication-with-logdelegate.cs"